# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: bitPOS - Bitcoin Lightning Wallet API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /entities:
    post:
      operationId: createEntity
      tags: [entities]
      summary: Create a new entity and account (signup)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Entity created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email or handle already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /entities/{id}:
    get:
      operationId: getEntity
      tags: [entities]
      summary: Get entity and account info
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Entity info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /entities/{id}/business:
    post:
      operationId: activateEntityBusiness
      tags: [entities]
      summary: Activate business profile for an entity's account
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessInput"
      responses:
        "200":
          description: Business profile activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountInfo"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    RegisterInput:
      type: object
      required: [email, handle, pin]
      properties:
        email:
          type: string
          format: email
        handle:
          type: string
          minLength: 3
          maxLength: 32
          pattern: "^[a-z0-9_]+$"
        pin:
          type: string
          minLength: 4
          maxLength: 8
...
    EntityInfo:
      type: object
      required: [id, email, handle, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        handle:
          type: string
        phone:
          type: ["string", "null"]
        phoneVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AccountInfo:
      type: object
      required: [id, type, balanceSats, businessActive]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [personal, business]
        businessName:
          type: ["string", "null"]
        businessActive:
          type: boolean
        balanceSats:
          type: integer
    AuthResponse:
      type: object
      required: [token, entity, account]
      properties:
        token:
          type: string
        entity:
          $ref: "#/components/schemas/EntityInfo"
        account:
          $ref: "#/components/schemas/AccountInfo"
    MeResponse:
      type: object
      required: [entity, account]
      properties:
        entity:
          $ref: "#/components/schemas/EntityInfo"
        account:
          $ref: "#/components/schemas/AccountInfo"
...
    BusinessInput:
      type: object
      required: [businessName]
      properties:
        businessName:
          type: string
          minLength: 1
          maxLength: 100
...