# lib/api-zod/src/generated/types/accountInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * bitPOS - Bitcoin Lightning Wallet API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountInfoType } from "./accountInfoType";

export interface AccountInfo {
  id: string;
  type: AccountInfoType;
  /** @nullable */
  businessName?: string | null;
  businessActive: boolean;
  balanceSats: number;
}