# lib/api-zod/src/generated/types/meResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * bitPOS - Bitcoin Lightning Wallet API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountInfo } from "./accountInfo";
import type { EntityInfo } from "./entityInfo";

export interface MeResponse {
  entity: EntityInfo;
  account: AccountInfo;
}