# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * bitPOS - Bitcoin Lightning Wallet API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  email: string;
  /**
   * @minLength 3
   * @maxLength 32
   * @pattern ^[a-z0-9_]+$
   */
  handle: string;
  /**
   * @minLength 4
   * @maxLength 8
   */
  pin: string;
}